requirements.txt
streamlit
ifcopenshell
fpdf
matplotlib
google-generativeai
numpy